# docs/swagger.yaml
definitions:
  models.Invoice:
    properties:
      cost:
        example: 1.56
        type: number
      ethAddress:
        example: a 42-char eth address with 0x prefix
        maxLength: 42
        minLength: 42
        type: string
    required:
    - ethAddress
    type: object
...
  routes.DownloadFileObj:
    properties:
      fileID:
        example: the handle of the file
        type: string
    required:
    - fileID
    type: object
...
  routes.StatusRes:
    properties:
      status:
        example: status of the request
        type: string
    type: object
  routes.UploadFileReq:
    properties:
      chunkData:
        example: a binary string of the chunk data
        type: string
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - chunkData
    - publicKey
    - requestBody
    - signature
    type: object
  routes.UploadFileSiaReq:
    properties:
      fileData:
        example: a binary string of the file data
        type: string
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - fileData
    - publicKey
    - requestBody
    - signature
    type: object
  routes.UploadStatusReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
...
  routes.accountCreateRes:
    properties:
      expirationDate:
        type: string
      invoice:
        $ref: '#/definitions/models.Invoice'
    required:
    - expirationDate
    type: object
...
  routes.createStripePaymentReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
...
  routes.downloadPublicFileRes:
    properties:
      fileDownloadThumbnailUrl:
        example: a URL to use to download the public file thumbnail
        type: string
      fileDownloadUrl:
        example: a URL to use to download the public file
        type: string
    type: object
...
  routes.stripeDataRes:
    properties:
      amount:
        type: number
      chargeID:
        type: string
      chargePaid:
        type: boolean
      opctTxStatus:
        type: string
      status:
        example: status of the request
        type: string
      stripePaymentExists:
        type: boolean
      stripeToken:
        type: string
    type: object
...
info:
  contact:
    name: Opacity Staff
    url: https://telegram.me/opacitystorage
  description: Opacity backend for file storage.
  license:
    name: OPACITY LIMITED CODE REVIEW LICENSE LICENSE.md
  termsOfService: https://opacity.io/terms-of-service
  title: Storage Node
  version: "1.0"
paths:
...
  /api/v1/stripe/create:
    post:
      consumes:
      - application/json
      description: |-
        create a stripe payment
        requestBody should be a stringified version of (values are just examples):
        {
        "stripeToken": "tok_KPte7942xySKBKyrBu11yEpf",
        "timestamp": 1659325302,
        }
      parameters:
      - description: stripe payment creation object
        in: body
        name: createStripePaymentReq
        required: true
        schema:
          $ref: '#/definitions/routes.createStripePaymentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.stripeDataRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: account is already paid for
          schema:
            type: string
        "404":
          description: 'no account with that id: (with your accountID)'
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: create a stripe payment
...
  /api/v1/upload:
    post:
      consumes:
      - multipart/form-data
      description: |-
        upload a chunk of a file. The first partIndex must be 1.
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "a deterministically created file handle",
        "partIndex": 1,
        }
      parameters:
      - description: an object to upload a chunk of a file
        in: body
        name: UploadFileReq
        required: true
        schema:
          $ref: '#/definitions/routes.UploadFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/routes.accountCreateRes'
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: upload a chunk of a file
  /api/v1/upload-status:
    post:
      consumes:
      - application/json
      description: |-
        check status of an upload
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "a deterministically created file handle",
        }
      parameters:
      - description: an object to poll upload status
        in: body
        name: UploadStatusReq
        required: true
        schema:
          $ref: '#/definitions/routes.UploadStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: signature did not match
          schema:
            type: string
        "404":
          description: file or account not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: check status of an upload
...
  /api/v2/download/public:
    post:
      consumes:
      - application/json
      description: returns the URLs for a public file and it's thumbnail, if no thumbnail
        is present, return a default one
      parameters:
      - description: download object for non-signed requests
        in: body
        name: routes.DownloadFileObj
        required: true
        schema:
          $ref: '#/definitions/routes.DownloadFileObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.downloadPublicFileRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "404":
          description: such data does not exist
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: returns the URLs for a public file and it's thumbnail
...
  /api/v2/sia/upload:
    post:
      consumes:
      - multipart/form-data
      description: |-
        upload a Sia file via a form
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "a deterministically created file handle",
        }
      parameters:
      - description: an object to upload a Sia file
        in: body
        name: UploadFileSiaReq
        required: true
        schema:
          $ref: '#/definitions/routes.UploadFileSiaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: signature did not match
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: upload a Sia file
  /api/v2/sia/upload-status:
    post:
      consumes:
      - application/json
      description: |-
        check status of a Sia upload
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "a deterministically created file handle",
        }
      parameters:
      - description: an object to poll upload status for a Sia file
        in: body
        name: UploadStatusReq
        required: true
        schema:
          $ref: '#/definitions/routes.UploadStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: signature did not match
          schema:
            type: string
        "404":
          description: file or account not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: check status of a Sia upload
...
  /api/v2/upload-status-public:
    post:
      consumes:
      - application/json
      description: |-
        check status of a public upload and creates a thumbnail in case the file is an image. If the mimeType is send, the API will not create the thumbnail
        "jpg" (or "jpeg"), "png", "gif", "tif" (or "tiff") and "bmp" are supported
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "a deterministically created file handle",
        "mimeType": "the mime type of the file",
        "title": "file title",
        "description": "a description to be used as metatags value"
        }
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: signature did not match
          schema:
            type: string
        "404":
          description: file not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: check status of a public upload
...
swagger: "2.0"